+++ core/include/detray/core/detail/container_views.hpp
@@ -8,13 +8,29 @@
 #pragma once
 
 // Project include(s)
+#include "detray/definitions/containers.hpp"
 #include "detray/definitions/qualifiers.hpp"
+
+// Vecmem include(s)
+#include <vecmem/containers/device_vector.hpp>
+#include <vecmem/containers/jagged_device_vector.hpp>
+
+// Thrust include(s)
+#include <thrust/tuple.h>
 
 // System include(s)
 #include <tuple>
 #include <type_traits>
 
 namespace detray {
+
+/// Container types used in device code
+using device_container_types =
+    container_types<vecmem::device_vector, thrust::tuple, darray,
+                    vecmem::jagged_device_vector>;
+
+/// How to obtain views for vecmem types
+using vecmem::get_data;
 
 namespace detail {
 
@@ -94,5 +110,81 @@ using get_view_t = typename get_view<T>::type;
 template <typename... view_ts>
 using dmulti_view = detray::detail::dmulti_view_helper<
     std::conjunction_v<detail::is_device_view<view_ts>...>, view_ts...>;
+
+/// @brief Detray version of 'get_data' - non-const
+///
+/// It is available to the generic containers before the value types are known,
+/// thus enabling 'duck typing'.
+///
+/// @note This does not pick up the vecmem types.
+template <class T,
+          std::enable_if_t<detail::is_device_view_v<typename T::view_type>,
+                           bool> = true>
+typename T::view_type get_data(T& viewable) {
+    return viewable.get_data();
+}
+
+/// @brief Detray version of 'get_data' - const
+///
+/// It is available to the generic containers before the value types are known,
+/// thus enabling 'duck typing'.
+///
+/// @note This does not pick up the vecmem types.
+template <class T,
+          std::enable_if_t<detail::is_device_view_v<typename T::view_type>,
+                           bool> = true>
+typename T::const_view_type get_data(const T& viewable) {
+    return viewable.get_data();
+}
+
+/// Type trait specializations for vecmem containers
+/// @{
+
+/// Specialized view for @c vecmem::vector containers
+template <typename T>
+using dvector_view = vecmem::data::vector_view<T>;
+
+/// Specialization of 'is view' for @c vecmem::data::vector_view containers
+template <typename T>
+struct detail::is_device_view<vecmem::data::vector_view<T>, void>
+    : public std::true_type {};
+
+/// Specialization of the view getter for @c vecmem::vector
+template <typename T>
+struct detail::get_view<vecmem::vector<T>, void> : public std::true_type {
+    using type = dvector_view<T>;
+};
+
+/// Specialization of the view getter for @c vecmem::vector
+template <typename T>
+struct detail::get_view<const vecmem::vector<T>, void> : public std::true_type {
+    using type = dvector_view<const T>;
+};
+
+/// Specialized view for @c vecmem::jagged_vector containers
+template <typename T>
+using djagged_vector_view = vecmem::data::jagged_vector_view<T>;
+
+/// Specialization of 'is view' for @c vecmem::data::jagged_vector_view
+/// containers
+template <typename T>
+struct detail::is_device_view<vecmem::data::jagged_vector_view<T>, void>
+    : public std::true_type {};
+
+/// Specialization of the view getter for @c vecmem::jagged_vector
+template <typename T>
+struct detail::get_view<vecmem::jagged_vector<T>, void>
+    : public std::true_type {
+    using type = djagged_vector_view<T>;
+};
+
+/// Specialization of the view getter for @c vecmem::jagged_vector
+template <typename T>
+struct detail::get_view<const vecmem::jagged_vector<T>, void>
+    : public std::true_type {
+    using type = djagged_vector_view<const T>;
+};
+
+/// @}
 
 }  // namespace detray
