+++ core/include/detray/definitions/containers.hpp
@@ -8,18 +8,11 @@
 #pragma once
 
 // Project include(s)
-#include "detray/core/detail/container_views.hpp"
 #include "detray/definitions/qualifiers.hpp"
-#include "detray/utils/type_traits.hpp"
 
 // Vecmem include(s)
-#include <vecmem/containers/device_vector.hpp>
-#include <vecmem/containers/jagged_device_vector.hpp>
 #include <vecmem/containers/jagged_vector.hpp>
 #include <vecmem/containers/vector.hpp>
-
-// Thrust include(s)
-#include <thrust/tuple.h>
 
 // System include(s)
 #include <array>
@@ -29,6 +22,7 @@
 #include <vector>
 
 namespace detray {
+
 template <typename value_t, std::size_t kDIM>
 using darray = std::array<value_t, kDIM>;
 
@@ -69,87 +63,49 @@ struct container_types {
 
 /// Defining some common types
 using host_container_types = container_types<>;
-using device_container_types =
-    container_types<vecmem::device_vector, thrust::tuple, darray,
-                    vecmem::jagged_device_vector>;
 
-/// How to obtain views for vecmem types
-using vecmem::get_data;
+namespace detail {
 
-/// @brief Detray version of 'get_data' - non-const
-///
-/// It is available to the generic containers before the value types are known,
-/// thus enabling 'duck typing'.
-///
-/// @note This does not pick up the vecmem types.
-template <class T,
-          std::enable_if_t<detail::is_device_view_v<typename T::view_type>,
-                           bool> = true>
-typename T::view_type get_data(T& viewable) {
-    return viewable.get_data();
-}
+// make std::get available in detray detail namespace, where also the thrust and
+// index specific overloads live.
+using std::get;
 
-/// @brief Detray version of 'get_data' - const
-///
-/// It is available to the generic containers before the value types are known,
-/// thus enabling 'duck typing'.
-///
-/// @note This does not pick up the vecmem types.
-template <class T,
-          std::enable_if_t<detail::is_device_view_v<typename T::view_type>,
-                           bool> = true>
-typename T::const_view_type get_data(const T& viewable) {
-    return viewable.get_data();
-}
+/// Trait class to figure out if a given type has a @c reserve(...) function
+template <typename T>
+struct has_reserve {
 
-/// Type trait specializations for vecmem containers
+    private:
+    /// Function returning @c std::true_type for types that do have a @c
+    /// reserve(...) function
+    template <typename C>
+    static constexpr auto check(C*) ->
+        typename std::is_void<decltype(std::declval<C>().reserve(
+            std::declval<typename C::size_type>()))>::type;
+    /// Function returning @c std::false_type for types that fair the previous
+    /// function
+    template <typename>
+    static constexpr std::false_type check(...);
+
+    /// Declare the value type of this trait class
+    typedef decltype(check<T>(nullptr)) type;
+
+    public:
+    /// Value of the check
+    static constexpr bool value = type::value;
+};
+
+/// @name Functions calling or not calling reserve(...) based on whether it's
+/// available
 /// @{
+template <typename T, std::enable_if_t<has_reserve<T>::value, bool> = true>
+DETRAY_HOST_DEVICE void call_reserve(T& obj, std::size_t newsize) {
+    obj.reserve(newsize);
+}
 
-/// Specialized view for @c vecmem::vector containers
-template <typename T>
-using dvector_view = vecmem::data::vector_view<T>;
-
-/// Specialization of 'is view' for @c vecmem::data::vector_view containers
-template <typename T>
-struct detail::is_device_view<vecmem::data::vector_view<T>, void>
-    : public std::true_type {};
-
-/// Specialization of the view getter for @c vecmem::vector
-template <typename T>
-struct detail::get_view<vecmem::vector<T>, void> : public std::true_type {
-    using type = vecmem::data::vector_view<T>;
-};
-
-/// Specialization of the view getter for @c vecmem::vector
-template <typename T>
-struct detail::get_view<const vecmem::vector<T>, void> : public std::true_type {
-    using type = dvector_view<const T>;
-};
-
-/// Specialized view for @c vecmem::jagged_vector containers
-template <typename T>
-using djagged_vector_view = vecmem::data::jagged_vector_view<T>;
-
-/// Specialization of 'is view' for @c vecmem::data::jagged_vector_view
-/// containers
-template <typename T>
-struct detail::is_device_view<vecmem::data::jagged_vector_view<T>, void>
-    : public std::true_type {};
-
-/// Specialization of the view getter for @c vecmem::jagged_vector
-template <typename T>
-struct detail::get_view<vecmem::jagged_vector<T>, void>
-    : public std::true_type {
-    using type = djagged_vector_view<T>;
-};
-
-/// Specialization of the view getter for @c vecmem::jagged_vector
-template <typename T>
-struct detail::get_view<const vecmem::jagged_vector<T>, void>
-    : public std::true_type {
-    using type = djagged_vector_view<const T>;
-};
-
+template <typename T, std::enable_if_t<!(has_reserve<T>::value), bool> = true>
+DETRAY_HOST_DEVICE void call_reserve(T& /*obj*/, std::size_t /*newsize*/) {}
 /// @}
+
+}  // namespace detail
 
 }  // namespace detray
