+++ core/include/detray/geometry/detector_volume.hpp
@@ -8,7 +8,6 @@
 
 // Project include(s)
 #include "detray/definitions/containers.hpp"
-#include "detray/definitions/detail/accessor.hpp"
 #include "detray/definitions/geometry.hpp"
 #include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
