+++ core/include/detray/geometry/volume_connector.hpp
@@ -8,7 +8,8 @@
 #pragma once
 
 // Project include(s)
-#include "detray/definitions/detail/accessor.hpp"
+#include "detray/definitions/containers.hpp"
+#include "detray/definitions/indexing.hpp"
 #include "detray/materials/predefined_materials.hpp"
 
 // Vecmem include(s)
