+++ core/include/detray/utils/ranges/iota.hpp
@@ -7,7 +7,7 @@
 #pragma once
 
 // Project include(s)
-#include "detray/definitions/detail/accessor.hpp"
+#include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
 #include "detray/utils/ranges/ranges.hpp"
 #include "detray/utils/type_traits.hpp"
