+++ core/include/detray/utils/type_registry.hpp
@@ -8,9 +8,9 @@
 #pragma once
 
 // Project include(s)
-#include "detray/definitions/detail/accessor.hpp"
 #include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
+#include "detray/utils/tuple_helpers.hpp"
 
 // System include(s)
 #include <type_traits>
