+++ core/include/detray/utils/type_traits.hpp
@@ -8,7 +8,7 @@
 #pragma once
 
 // Project include(s)
-#include "detray/definitions/detail/accessor.hpp"
+#include "detray/definitions/indexing.hpp"
 
 // System include(s)
 #include <type_traits>
@@ -82,6 +82,52 @@ struct is_interval<
 
 template <typename TYPE>
 inline constexpr bool is_interval_v = is_interval<TYPE>::value;
+/// @}
+
+/// Extract the first type from a parameter pack.
+/// @{
+template <typename first_t = void, typename... other_t>
+struct first_type {
+    using type = first_t;
+};
+
+template <typename... TYPES>
+using first_t = typename first_type<TYPES...>::type;
+
+/// Extract the first value from an index pack.
+template <std::size_t first_t, std::size_t... other_t>
+struct first_idx {
+    static constexpr std::size_t value = first_t;
+};
+
+template <std::size_t... TYPES>
+inline constexpr std::size_t first_idx_v = first_idx<TYPES...>::value;
+/// @}
+
+/// Get the position of a type in a template parameter pack
+/// @{
+template <typename T, typename... Ts>
+struct get_type_pos {
+
+    /// Unroll a parameter pack without using a tuple.
+    ///
+    /// @note Returns the position of the type counted from the back!
+    template <typename first_t, typename... remaining_types>
+    DETRAY_HOST_DEVICE static inline constexpr std::size_t type_pos_back() {
+        if constexpr (not std::is_same_v<T, first_t>) {
+            return type_pos_back<remaining_types...>();
+        }
+        if constexpr (std::is_same_v<T, first_t>) {
+            return sizeof...(remaining_types) + 1;
+        }
+        return std::numeric_limits<std::size_t>::max();
+    }
+
+    static constexpr std::size_t value = sizeof...(Ts) - type_pos_back<Ts...>();
+};
+
+template <typename T, typename... Ts>
+inline constexpr std::size_t get_type_pos_v = get_type_pos<T, Ts...>::value;
 /// @}
 
 }  // namespace detray::detail
